[DEV-9662] refactor code
