+++ packages/chart/src/components/LineChart/index.tsx
@@ -14,7 +14,7 @@ import ConfigContext from '../../ConfigContext'
 import useRightAxis from '../../hooks/useRightAxis'
 
 // Local helpers and components
-import { filterCircles, createStyles, createDataSegments } from './helpers'
+import { filterCircles, createStyles, createDataSegments, truncateText } from './helpers'
 import LineChartCircle from './components/LineChart.Circle'
 import LineChartBumpCircle from './components/LineChart.BumpCircle'
 import isNumber from '@cdc/core/helpers/isNumber'
@@ -22,7 +22,6 @@ import isNumber from '@cdc/core/helpers/isNumber'
 // Types
 import { type ChartContext } from '../../types/ChartContext'
 import { type LineChartProps } from './LineChartProps'
-import { getTextWidth } from '@cdc/core/helpers/getTextWidth'
 
 const LineChart = (props: LineChartProps) => {
   // prettier-ignore
@@ -56,21 +55,6 @@ const LineChart = (props: LineChartProps) => {
 
   const xPos = d => {
     return xScale(getXAxisData(d)) + (xScale.bandwidth ? xScale.bandwidth() / 2 : 0)
-  }
-
-  const truncateText = (text, maxWidth) => {
-    let width = getTextWidth(text, `normal ${16}px sans-serif`)
-    if (width <= maxWidth) {
-      return text // If within maximum width, return full text
-    }
-
-    // Truncate text by progressively removing characters until it fits
-    while (width > maxWidth && text.length > 0) {
-      text = text.slice(0, -1)
-      width = getTextWidth(text + '...')
-    }
-
-    return text + '...'
   }
 
   return (
